backend/requirements.txt
Flask>=3.0,<4.0
SQLAlchemy>=2.0,<3.0
requests>=2.32
Pillow>=10.0
python-dotenv>=1.0